feat(front): add obj viewer + add xyz viewer + refactor 3d viewer in a single component
