+++ frontend_app_file/src/component/ThreedViewer/ThreeDViewer.styl
@@ -1,3 +1,3 @@
-.PointCloudViewer
+.ThreeDViewer
   width 100%
   height 100%
